rename dcl_user_profile.rs fix typo
