+++ rust/decentraland-godot-lib/src/avatars/mod.rs
@@ -1,4 +1,4 @@
 pub mod avatar_scene;
 pub mod avatar_type;
-pub mod dcl_godot_profile;
+pub mod dcl_user_profile;
 pub mod wearable;
